Add signed int64 bit logging as a decimal number
Switched uint64_t casting to int64_t casting and added a few more
int casting to int64_t casting for more edge case like large LBAs
and such.

Switch custom int and int64_t conversion in `ZuluIDE_log.cpp` to
using the C++ standard library to_chars() function. This covers
the edge case when printing a largest negative number for the integer
type. This is most likely unnecessary as there currently isn't
any printing of negative numbers that would get that low and was only
noticed when trying to print the limits for testing.

+++ lib/ZuluIDE_platform_RP2350/rp2350_sniffer.cpp
@@ -307,7 +307,7 @@ void rp2350_sniffer_poll()
 
     if (!g_sniffer.should_sync && (uint32_t)(millis() - g_sniffer.sync_time) > SNIFFER_SYNC_INTERVAL)
     {
-        logmsg("-- Bus sniffer status: total ", (int)((g_sniffer.total_bytes + 1023) / 1024), " kB, ",
+        logmsg("-- Bus sniffer status: total ", (int64_t)((g_sniffer.total_bytes + 1023) / 1024), " kB, ",
                 (int)g_sniffer.overruns, " buffer overruns");
 
         g_sniffer.should_sync = true;
